# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IResponse": {
			"properties": {
				"code": {
					"type": "string"
				},
				"data": {
					"additionalProperties": true,
					"type": "object"
				}
			},
			"required": [
				"code",
				"data"
			],
			"type": "object",
			"additionalProperties": true
		},
		"IErrorResponse": {
			"properties": {
				"status": {
					"type": "number",
					"format": "double"
				},
				"code": {
					"type": "string"
				},
				"data": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"status",
				"code"
			],
			"type": "object",
			"additionalProperties": true
		},
...
		"IUpdateUserRequest": {
			"properties": {
				"nickName": {
					"type": "string"
				},
				"gender": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"city": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"mobile": {
					"type": "string"
				},
				"wechatId": {
					"type": "string"
				},
				"company": {
					"type": "string"
				},
				"avatarImage": {
					"type": "string"
				},
				"ageTag": {
					"items": {
						"type": "string"
					},
					"type": "array"
				}
			},
			"required": [
				"nickName",
				"gender",
				"description",
				"city",
				"email",
				"mobile",
				"wechatId",
				"company",
				"avatarImage",
				"ageTag"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "jubenfan-server",
		"version": "3.0.0",
		"description": "API description can go here",
		"license": {
			"name": "MIT"
		}
	},
	"paths": {
...
		"/users/{userId}": {
			"get": {
				"operationId": "getUserDetailById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"access_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "updateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/IErrorResponse"
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"access_token": [
							"user:updateUser"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUpdateUserRequest"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header",
			"template": "Bearer {apiKey}"
		}
	},
	"host": "localhost:3000",
	"schemes": [
		"http",
		"https"
	]
}